third_party/WebKit/LayoutTests/media/media-player-request-fullscreen.html

<!doctype html>
<html>
    <head>
        <title>Test mediaPlayerRequestFullscreen() without a user gesture.</title>
        <script src=media-file.js></script>
        <script src=video-test.js></script>
    </head>
    <body>
        <p>Test mediaPlayerRequestFullscreen() without a user gesture.</p>
        <video controls></video>
        <script>
            findMediaElement();
            video.src = findMediaFile("video", "content/test");
            waitForEvent('loadedmetadata', function()
            {
                internals.mediaPlayerRequestFullscreen(video);
                waitForEvent('webkitfullscreenchange', endTest);
                waitForEventAndFail('webkitfullscreenerror');
            });
        </script>
    </body>
</html>
