third_party/WebKit/LayoutTests/media/media-player-request-fullscreen-expected.txt

Test mediaPlayerRequestFullscreen() without a user gesture.

EVENT(loadedmetadata)
EVENT(webkitfullscreenchange)
END OF TEST
